Train/requirements.txt
torch==2.4.1
torchvision==0.19.1
numpy>=1.23,<2.0
pandas>=1.5
scikit-learn>=1.3
opencv-python>=4.8
Pillow>=10.0
matplotlib>=3.7
tqdm>=4.66
PyYAML>=6.0
streamlit>=1.35
fastapi>=0.111
uvicorn[standard]>=0.30